Add files via upload
I am the type of person that loves listen to the same song once i fall in love with it. 
I have spotify but when i am work, i listen youtube to listen to music
But you youtube does not have a replay button! 
So i just stop my work to replay my current fav song.
Not anymore! 
I wrote a small script that will replay the my song for me when its coming to an end. 
MY own replay button i guess !